server/requirements.txt
Click==7.0
fastapi==0.68.0
h11==0.9.0
httptools==0.0.13
psycopg2-binary==2.8.4
pydantic>=1.6.2,<2.0.0
SQLAlchemy==1.3.12
starlette==0.14.2
typing==3.7.4.1
uvicorn==0.11.1
uvloop==0.14.0
websockets==8.1
PyJWT
